Implement custom type generation
Error types are just special case of custom types. I adjusted the code
generation to treat both types pretty much the same. Any custom type has an
arbitrary `Object` payload which allows for generalization. That also seems to
be the way the RPC spec handles theses types, so this approach should be golden.