The tests were actually not working properly with the last implementation of the plugin registration
The lazy loading was quite a stupid idea because you were not able to call the
functions without prior calling them. Now, the functions are registered as soon
as the provider has started up.

There is still some caveat. For the plugins to work, one has to make any
`rpcrequest` to the host, so that the functions are definitely all loaded.
Since we ship some functions to do so, the startup code must be adjusted.