Implement restarting of the plugin provider
An MVar is used to signal the main thread that some kind of quit action should
be performed. Firstly I tried to achieve that via exception handling, but the
exception did not propagate to the main thread for some reason and it made the
code blow up a bit. Implementing the MVar-solution was very painless and
quickly done in comparison. There might be a better way, but local testing
indicates that it should work.