Let nvim-hs handle the sandbox environment variables
If nvim-hs is started from within a sandbox srcipt, some environment variables
are set so that other tools (e.g. ide-backend or ghcid) cannot operate properly
without (un-)setting them to ther needs. To mitigate this for most plugin
authors, nvim-hs now saves the state for this variables and only sets them for
the time it is recompiling.