The startup code now runs in an environment with extra information
A data type `StartupConf` has been added that contains more or less
useful information for plugins about some of the internals of nvim-hs.
This data type can be extended in the future without breaking any code,
so this should mark the point at which the plugin startup code becomes
backwards compatible.