Remove clean up of autocmd
The cleaning should remove registered autocommands, functions and
commands from plugins when nvim-hs is restarted. It was never
implemented for fucntions and commands and now the code for the
autocmnds no longer typechecks. Maybe it is simpler to remove the code
altogether and just force a restart of nvim to clean things up.