Remove mechanism suppressing exceptions from RPC calls
Instead of calling error with an unhelpful message, simply throw the NeovimException.
The method that has thrown the excpetion has been added to the exception object.

A helper function `catchNeovimException` has been added which is simply
a specialization of the normal `catch`.

The unhelpful error message was a remnant of a refactor which removed
error types from the remote functions because all remote functions can fail.